refactor: restructure HeroSection(i.e, remove conditional rendering) and SideMenu components; extract constants for menu and social links,
