+++ src/components/features/HeroSection/SideMenu.tsx
@@ -1,67 +1,15 @@
 import Link from "next/link";
-import discord1 from "@/assets/discord-icon-purp.svg";
-import github from "@/assets/github-mark.svg";
-import browser from "@/assets/browser-icon.svg";
-import x from "@/assets/x-icon.svg";
+
 import { cn } from "@/util/cn";
 import { MenuItem } from "../../ui/navigation/MenuItem";
 import { SocialLinks } from "@/components/ui/navigation/SocialLinks";
 import { ExitButton } from "@/components/ui/buttons/ExitButton";
+import { MENU_ITEMS } from "./constants";
+import { SOCIAL_LINKS } from "./constants";
 interface SideMenuProps {
 	open: boolean;
 	setOpen: React.Dispatch<React.SetStateAction<boolean>>;
 }
-
-interface MenuItem {
-	label: string;
-	href: string;
-}
-
-interface SocialLink {
-	icon: string;
-	href: string;
-	alt: string;
-	width: number;
-	height: number;
-}
-
-const MENU_ITEMS: MenuItem[] = [
-	{ label: "Home", href: "/" },
-	{ label: "About us", href: "#about" },
-	{ label: "Team", href: "#team" },
-	{ label: "Testimonials", href: "#testimonials" },
-] as const;
-
-const SOCIAL_LINKS: SocialLink[] = [
-	{
-		icon: discord1,
-		href: "#",
-		alt: "Discord icon",
-		width: 20,
-		height: 20,
-	},
-	{
-		icon: github,
-		href: "#",
-		alt: "Github icon",
-		width: 20,
-		height: 20,
-	},
-	{
-		icon: browser,
-		href: "#",
-		alt: "Browser icon",
-		width: 20,
-		height: 20,
-	},
-	{
-		icon: x,
-		href: "#",
-		alt: "X icon",
-		width: 20,
-		height: 20,
-	},
-] as const;
 
 export default function SideMenu({ open, setOpen }: SideMenuProps) {
 	return (
